Lots of style changes, fix tooltips, begin implementing migration for manga

+++ database/database_structure.sql
@@ -22,7 +22,7 @@ CREATE TABLE `episode` (
   `id` int(11) NOT NULL,
   `series_id` int(11) NOT NULL,
   `season_id` int(11) DEFAULT NULL,
-  `number` int(11) DEFAULT NULL,
+  `number` decimal(10,2) DEFAULT NULL,
   `name` varchar(200) DEFAULT NULL,
   `duration` int(11) DEFAULT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
